extmod/moductypes: Add support for union fields for .calc_offsets().
A field which should share storage with the previous one, should have
uctypes.PREV_OFFSET value used as the offset. In other words, every
field but the first of the union should use uctypes.PREV_OFFSET. This
allows to handle both anonymous unions (when union fields are at the
same level as other fields) and classical C unions, when union fieldset
is referred by a name.