py/objarray: Introduce "memview_offset" alias for "free" field of the object.
Both mp_type_array and mp_type_memoryview use the same obejct structure,
mp_obj_array_t, but for the case of memoryview, some fields, e.g. "free",
have different meaning. As the "free" field is also a bitfield, assume
that (anonymous) union can't be uses here, and just a field alieas using
a #define. As it's a define, it should be a selective identifier, so use
verbose "memview_offset" to avoid any clashes.