py/gc.h: Don't depend on any uPy headers.
GC implementation is the lowest layer of MicroPython, everything else is
built on top of it. For example, memory allocation (misc.h) is done via
GC implementation, so there's no need to include misc.h into gc.h (but
may be needed to do the opposite, so avoid circular dependency). Actually,
gc.h (header, not implementation) doesn't even depend on mpconfig.h, so
remove that either.