docs/reference/glossary: Describe callee-owned tuple as a tuple subtype.
This allows to fully incorporate callee-owned tuple to the language
semantics: if it's a subtype, surely it can acquire additional behavior
(mutability by special functions) comparing to the main (immutable) tuple
type. A not is made that current implementation doesn't make the technical
distinction for efficiency reason (i.e. it's implied subtype).