extmod/moduselect: poll.unregister: Add optional throw=True param.
If 2nd, positional only argument of False is passed to poll.unregister()
method, then it won't throw exception if passed a stream object not
registered with the poller, which was the previous behavior. Then default
behavior is made compatible with CPython's, which is to throw exception
in this case.

This is similar to the change previously done to the Unix port.

Change-Id: If2c25038a2fe1dd75acd02e3305beecf7e868de9