py/objmodule: Implement __setattr__ for native modules.
This can be seen as a counterpart to PEP562's module-level __getattr__.
Except that it applies only to native modules, as these don't have
other means to override attribute's value. (On the other hand, for
Python-level modules, assignment to any attribute will update its value,
or even create it if it didn't exist before).

Change-Id: I64d92439a7af157ec14acecaae12ac9fab9c59ee